+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.11] - 2025-07-31
+
+### Fixed
+- Fixed MAAS API DNS record creation by using FQDN field directly instead of splitting into name and domain
+- Removed trailing dots from FQDNs as MAAS API does not expect them
+- Removed obsolete recordName function that was causing empty name parameters
+
 ## [1.0.10] - 2025-07-31
 
 ### Fixed